+++ packages/conc/changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.4.0.0
+
 * Add `lock`, a combinator for protecting a region with a mutex.
 * Add `scheduleAsync`, a combinator for running an action async that allows the handle to be used before the thread
   starts